refactor(swm): share peer identity and split fanout tests
