+++ packages/core/src/index.ts
@@ -163,6 +163,9 @@ export {
   type DialOpts,
   type ProtocolHandler,
   LibP2PNetwork,
+  canonicalPeerIdString,
+  tryCanonicalPeerIdString,
+  type CanonicalPeerId,
   type NetworkStateRegistry,
   StubNetworkStateRegistry,
   type AgentDirectoryLookup,
